pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apikeyrotator"
version = "0.2.0"
description = "Ultra simple API key rotation for bypassing rate limits"
readme = "README.md"
authors = [
    { name = "Prime Evolution", email = "develop@eclips-team.ru" }
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["api", "rotation", "http"]
dependencies = ["requests>=2.25.0", "aiohttp>=3.8.0", "python-dotenv>=1.0.0", "PyYAML>=6.0", "boto3>=1.28.0"]

[project.urls]
Homepage = "https://github.com/PrimeevolutionZ/apikeyrotator"
Repository = "https://github.com/PrimeevolutionZ/apikeyrotator"
Issues = "https://github.com/PrimeevolutionZ/apikeyrotator/issues"
